# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Daftar Properti API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tele-app/users/profile": {
            "get": {
                "tags": [
                    "Telegram Users"
                ],
                "summary": "Get profile",
                "description": "Returns user profile",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TelegramUserProfile"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Telegram Users"
                ],
                "summary": "Update profile",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TelegramUserProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TelegramUserProfile"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TelegramUserProfileRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Jono Doe"
                    },
                    "phoneNumber": {
                        "type": "string",
                        "example": "081111111111"
                    },
                    "city": {
                        "type": "string",
                        "example": "Surabaya"
                    },
                    "description": {
                        "type": "string",
                        "example": "Agen terpercaya"
                    },
                    "company": {
                        "type": "string",
                        "example": "Agen XXX"
                    },
                    "picture": {
                        "type": "string",
                        "format": "binary",
                        "example": "\u0000\u0000\u0000\u0004\u0000\u0000\u0000\u0004"
                    },
                    "isPublicProfile": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
...
            "TelegramUserProfile": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "123"
                    },
                    "publicId": {
                        "type": "string",
                        "example": "id-123"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phoneNumber": {
                        "type": "string",
                        "example": "0811111"
                    },
                    "city": {
                        "type": "string",
                        "example": "New York"
                    },
                    "description": {
                        "type": "string",
                        "example": "I am a programmer"
                    },
                    "picture": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "company": {
                        "type": "string",
                        "example": "Google"
                    },
                    "isPublicProfile": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}